requirements/base.txt
django==1.10.5
ipython==5.0.0
raven==5.24.3

psycopg2==2.6.2
PyYAML==3.11
redis==2.10.5
certifi==2016.8.8
click==6.6
invoke==0.13.0
celery==3.1.23

djangorestframework==3.4.3
django-extensions==1.7.0
django-cors-headers==1.1.0
django-oauth-toolkit==0.10.0
django-basis==0.4.0
django-environ==0.4.0
django-redis==4.4.4
django-filter==0.13.0
django-debug-toolbar==1.6

elasticsearch==2.3.0
